+++ freeproxy/modules/proxies/kuaidaili.py
@@ -8,7 +8,7 @@ WeChat Official Account (微信公众号):
 '''
 import re
 import requests
-from user_agent import generate_user_agent
+from fake_useragent import UserAgent
 try:
     from base import BaseProxiedSession
 except:
@@ -26,7 +26,7 @@ class KuaidailiProxiedSession(BaseProxiedSession):
         # obtain proxies: 'https://www.kuaidaili.com/free/inha/1/'
         for page in range(1, self.max_pages+1):
             url = f'https://www.kuaidaili.com/free/inha/{page}/'
-            headers = {'User-Agent': generate_user_agent()}
+            headers = {'User-Agent': UserAgent().random}
             resp = requests.get(url, headers=headers)
             if resp.status_code != 200: continue
             proxies = re.findall(r'const fpsList = (\[.*?\]);', resp.text)
@@ -40,7 +40,7 @@ class KuaidailiProxiedSession(BaseProxiedSession):
         # obtain proxies: 'https://www.kuaidaili.com/free/dps/1/'
         for page in range(1, self.max_pages+1):
             url = f'https://www.kuaidaili.com/free/dps/{page}/'
-            headers = {'User-Agent': generate_user_agent()}
+            headers = {'User-Agent': UserAgent().random}
             resp = requests.get(url, headers=headers)
             if resp.status_code != 200: continue
             proxies = re.findall(r'const fpsList = (\[.*?\]);', resp.text)
@@ -55,7 +55,7 @@ class KuaidailiProxiedSession(BaseProxiedSession):
         # obtain proxies: 'https://www.kuaidaili.com/free/intr/1/'
         for page in range(1, self.max_pages+1):
             url = f'https://www.kuaidaili.com/free/inha/{page}/'
-            headers = {'User-Agent': generate_user_agent()}
+            headers = {'User-Agent': UserAgent().random}
             resp = requests.get(url, headers=headers)
             if resp.status_code != 200: continue
             proxies = re.findall(r'const fpsList = (\[.*?\]);', resp.text)
@@ -69,7 +69,7 @@ class KuaidailiProxiedSession(BaseProxiedSession):
         # obtain proxies: 'https://www.kuaidaili.com/free/fps/1/'
         for page in range(1, self.max_pages+1):
             url = f'https://www.kuaidaili.com/free/fps/{page}/'
-            headers = {'User-Agent': generate_user_agent()}
+            headers = {'User-Agent': UserAgent().random}
             resp = requests.get(url, headers=headers)
             if resp.status_code != 200: continue
             proxies = re.findall(r'const fpsList = (\[.*?\]);', resp.text)