+++ freeproxy/modules/proxies/proxylistplus.py
@@ -8,7 +8,7 @@ WeChat Official Account (微信公众号):
 '''
 import requests
 from bs4 import BeautifulSoup
-from user_agent import generate_user_agent
+from fake_useragent import UserAgent
 try:
     from base import BaseProxiedSession
 except:
@@ -26,7 +26,7 @@ class ProxylistplusProxiedSession(BaseProxiedSession):
         # obtain proxies
         for page in range(1, self.max_pages+1):
             url = f'https://list.proxylistplus.com/Fresh-HTTP-Proxy-List-{page}'
-            headers = {'User-Agent': generate_user_agent()}
+            headers = {'User-Agent': UserAgent().random}
             resp = requests.get(url, headers=headers)
             if resp.status_code != 200: continue
             soup = BeautifulSoup(resp.text, 'lxml')